Migrate to vitest (#664)
* Convert all tests to vitest (some failures)

All tests in js/ and react/ pass.

Most tests in components/ pass, except for some of thes mocking tests,
but not all.

* Mock test updates partially complete

* Update  GitHub actions

* Update version.ts to 1.2.6 in js, react

* remove use of any when mocking fetch builtin

* Formatting-only change

* Remove instructions for eslint-plugin-jest

Moving to vitest means we are not using this.

* Adjust imports to vitest reqs; add dedicated test-only tsconfigs

* Centralize import of jest-dom in components/vitest.setup.ts

* Separate lint configs for tests vs prod in js/

react/ and components/ don't use typed linting. Perhaps they should, though.

* Avoid calling vi.mock() in skipped tests

The incomplete mocks were overriding the complete mock, causing the
error: No "useIsLightBackground" export is defined on the  "../../../hooks" mock

* Fix build

* Add a matrix of node versions

* Remove trailing comma

* fix missing eslint

* Remove v25 from components/

* Add mock localStorage

* Add (temporarily) a dedicated test:node25

* Use a wrapper test.sh script instead of multiple package.json scripts