fix: allow changing plan after plan stage skipped (#807)
Previously a useEffect in CheckoutDialog would not allow changes
to a selected plan, continually overwriting it with the value of
`checkoutState.planId`.

We can clear this state when navigating backwards after a deep link,
e.g. a checkout flow started from `initializeWithPlan`. Setting
`bypassPlanSelection: false` ensures that this process of clearing
the data only happens once.